Offset constants
First, the offsets are just harder to keep track of in an array.
You can't look at the code where they're used and know that you're
using the right one just by the array index.  So I've split them
out into separate constants, consistent with how the ports and
encoders are named.

Also, I've mathed out each offset value so they're just a numerical
offset, in degrees (and a conversion to radians).